cpp11-migrate: Transforms collect timing data.
Using updated form of newFrontendActionFactory(), Transforms now automatically
measure, if requested, how long it takes to apply a MatchFinder to a source
file. Other per-transform overhead, e.g. applying replacements, is not
currently measured. This behaviour is disabled for now and soon will be
connected to a new command line arg.



git-svn-id: https://llvm.org/svn/llvm-project/clang-tools-extra/trunk@182942 91177308-0d34-0410-b5e6-96231b3b80d8