cpp11-migrate: Add a class to support include directives modifications
The IncludeDirectives class helps with detecting and modifying #include
directives. For now it allows the users to add angled-includes in a source file.
This is a start for this class that will evolve in the future to add more
functionality.

This should fix the reverted commit r189037 (buildbot failures on Windows).

git-svn-id: https://llvm.org/svn/llvm-project/clang-tools-extra/trunk@189354 91177308-0d34-0410-b5e6-96231b3b80d8