[clang-tidy] Don't emit the same fixit multiple times.
If we had many header files we would attach the fix-it for all files to all
warnings, oops. This is harmless 99.9% of the time but can confuse the rewriter
in some edge cases. Sadly I failed to create a small test case for this.

While there move fix-its instead of copying.

git-svn-id: https://llvm.org/svn/llvm-project/clang-tools-extra/trunk@217951 91177308-0d34-0410-b5e6-96231b3b80d8