[clang-tidy] Checker for inaccurate use of erase() method.
Algorithms like remove() does not actually remove any element from the
container but returns an iterator to the first redundant element at the end
of the container. These redundant elements must be removed using the
erase() method. This check warns when not all of the elements will be
removed due to using an inappropriate overload.

Reviewer: alexfh

Differential Revision: http://reviews.llvm.org/D7496


git-svn-id: https://llvm.org/svn/llvm-project/clang-tools-extra/trunk@228679 91177308-0d34-0410-b5e6-96231b3b80d8